requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
